// src/components/presence-mouse/html/types.ts
import { Participant } from '../../../common/types/participant.types';

export interface ParticipantMouse extends Participant {
  slotIndex: number;
  x: number;
  y: number;
  visible: boolean;
}

export interface PresenceMouseProps {
  onGoToPresence?: (position: { x: number; y: number }, elementId: string) => void;
  dataAttributeName?: string;
}